requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors==0.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
install==1.3.5
pytz==2022.2.1
sqlparse==0.2.4
tzdata==2022.4
psycopg2==2.9.4